pollsapi/requirements.txt
# asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
drf-dynamic-fields==0.3.1
Pillow==8.1.2
drf-yasg==1.20.0
# pytz==2021.1
# sqlparse==0.4.1
